Misa-Cash/requirements.txt
# Análise de dados e Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1

# Visualização de dados
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Processamento de dados financeiros
yfinance==0.2.28
ta==0.10.2
pandas-ta==0.3.14b

# API e Web
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0

# Banco de dados
sqlalchemy==2.0.20
pymongo==4.4.1

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
joblib==1.3.2

# Testes
pytest==7.4.2
pytest-cov==4.1.0

# Documentação
sphinx==7.1.2
sphinx-rtd-theme==1.3.0 